Telephony: Specify file path when reading contacts from EF_ADN
Some cards, like RUIM and UICC with CSIM, have the same file ID for
multiple EFs. e.g. EF_ADN and EF_CSIM_LI. When trying to retrieve
the EF_ADN from such a card, the path gets mapped to that of
EF_CSIM_LI resulting in incorrect behavior.

To avoid this, specify the EF Id and the path to uniquely identify
EF_ADN, when reading contacts.
CRs-Fixed: 428259

Change-Id: I517590a378fe387b9fa8420e069ff4adc61cecb4