STK: Handle ICC refresh and card removal, and re-work clean up
Register for ICC refresh and let CatService's clients know when
an ICC refresh occurs.

Monitor card status. When card status moves to absent/present let
clients know.

Don't dispose CatService based on the number of applications.
Instead dispose it when UiccCard goes away

This helps avoid the race condition in the following scenario:
1. Radio is on and card is present
2. Turn Radio Off
3. Turn Radio On
4. Proactive command received

If CatService was disposed when card goes away on radio off
then it is possible for the unsol proactive command to be received
before Catservice is up and running on radio on.

Change-Id: I2d8d4eb1911c1cc1ed36620baa3508498f4c20b1
CRs-Fixed:270755, 283523, 450167, 442985