Telephony: Specify file path when reading contacts from EF_ADN
Some cards, like RUIM and UICC with CSIM, have the same file ID for
multiple EFs. e.g. EF_ADN and EF_CSIM_LI. When trying to retrieve
the EF_ADN from such a card, the path gets mapped to that of
EF_CSIM_LI resulting in incorrect behavior.
To avoid this, specify the EF Id and the path to uniquely identify
EF_ADN, when reading contacts.

CRs-Fixed: 504523

Change-Id: I8e96c5bbac333fb82add961d1d83826b679cbaea