Telephony: Add isRadioOn() method
The implementation of ITelephony.isRadioOn() method checks for
service state instead of radio state. In cases of basebandless
targets, service state is not changed and would remain in the
initialized OUT_OF_SERVICE state. This causes delay in shutdown
procedure for such targets.

Hence a method isRadioOn() is added in phone interface to get
the radio state.

Change-Id: I33760b710a1ef5ba50b92a014c72f33d8dc5e782
CRs-Fixed: 502317