Telephony: Check app type before checking the live network SID/NID
Does not check if the app type is CSIM or not before invoking
isInHomeSidNid() to check if the live network SID/NID is in the HOME
SID/NID list. We should not check it when the app type is not CSIM.

Check the app type first before invoking isInHomeSidNid() to check
the network SID/NID, if the app type is not the CSIM, we don't check it,
set the service provider name stored in SIM to the operator name.

CRs-Fixed: 546263

Change-Id: Iaa81a9492991f849d938f6562ef9d962c8b7abb7