telephony: Can not delete sim contacts
When delete a sim contacts with only one number, because
of only have number field, so when the program analysis
every field, other field's val is null except number field.
it will lead to a null pointer occur.

Reference MSimIccProvider, add the judgment for
array length, when the program analysis field, if the
length is not 2, continue to analysis next field.

CRs-Fixed: 564811

Change-Id: Ie3a3f3f26d4c25870fa94103ebff79e7c44a506a