Telephony: Fix cannot delete the sim contacts which contain anr or email
There may more than one PBR files in the USIM card, when we save an adn
record which contain anr or email to the sim card, there is not judgment
to check this PBR files has free space to save anr or email. So the anr
or email is not really saved to the USIM card.
Then we try delete this sim contact, the contact with anr or email can
not be matched in the sim, so it cannot be deleted.

Add the judgment to check this PBR files has free space to save anr or
email, if the current PBR file cannot save the AdnRecord which contain
anr or email, try save it into next PBR file.

CRs-Fixed: 568749

Change-Id: I590df4080b01afe74dc91177a4d7ba521f351e72