Settings: Fix the incorrect prompt of PIN when enable SUB1
The prompt displays incorrect when enable SUB1. And the remaining
attempts of PIN is the number of last time when enable the SUB1,
because the number of remaining attempts of PIN doesn't reset to
the initial number when the PIN of inputting is right.

Reset it to the initial number when the PIN of inputting is right

CRs-fixed: 570177

Change-Id: Iec809ae3d8a43058ed4f172b6d110f7fcabbf31b