Refactor IMSI and SubscriberId handling for CDMAPhone/CDMALTEPhone
IccRecords
CdmaServiceStateTracker does not need to set IMSI on an IccRecord.
CDMAPhone
When queried for the subscriber Id return the IMSI from
CdmaServiceStateTracker or the card depending on whether the active
subscription source is NV or RUIM respectively.
CDMALTEPhone
Don't return the IMSI from the SIM app (which is used for LTE) when
queried for the SubscriberId. Defer to the super class CDMAPhone's
implementation.
RuimRecords
- Override the getIMSI method.
- Remove our local copy of the IMSI.

Change-Id: Id613f2df91b94ae5d7613dfcef3e097cb817ff4e
CRs-Fixed: 323977,323977,342146,505580