Fix to avoid disconnected indication on call end
Voice call is in progress, and PS is reported as out of service
due to IRAT from 3G to 2G. On voice call end, as PS is out of
service, data is reported as disconnected to connectivityservice,
though the data connection is still active at modem.

Use the existing property "persist.telephony.oosisdc" to block
disconnected indication on voice call end.

Change-Id: If8d96115fdb8b0f65d75b57ee32da5fec2f4c178
CRs-Fixed: 636080