Uicc: Update icc card state when radio is turned on.
When the radio is off, if the icc card status changes,
the icc card state will be updated. At a later point when
the radio turns on, if the icc card status change event is not
received, ,the card state will not be restored and an incorrect
card state may be present in IccCardProxy.

Always update the icc card state (externalState) when the
radio turns on.

Change-Id: I187c9b27a799b9654ee65198c0e7053640d81e30
CRS-Fixed: 646808