Fix race condition issue in DcTracker enableApnType.
If DcTracker.disableApnType() and enableApnType() functions are called
back to back within short intervals, there is a possibility that the
ApnContext might still be in CONNECTED state while actually the ApnContext
is undergoing cleanup. In this case if enableApnType() is called,
it simply returns without enabling the ApnContext leaving it disabled
till next enable request comes.
To fix this, call setEnabled() always irrespective of the current
ApnContext state.

Change-Id: I7642cf4e439e0372bc0d7744c8dffb0af8110c58
CRs-Fixed: 699931