Fix for two messages inserted to sent table when sent via MAP.
SMSDispatcher writes the sent sms to db if the application which
is sending the sms is not default sms application. This wont
work for Bluetooth MAP profile.
Transparent is a parameter used by bluetooth MAP to determine
whether to write the messages to sent folder or not. If this
parameter is "ON", sms is not written to sent table of sms db.
If this parameter is "OFF", sms is written to sent table.
Based on the transparent parameter, the bluetooth app
can decide whether to write the sms to sent table or not.

Fix: If the package name is org.codeaurora.bluetooth, do not
write the sms to sent table.

Change-Id: I6367850173756e848c5462ffa35752ba4ba63412
CRs-Fixed: 703889