fix deadlock caused by back-to-back phonebook loading requests
There is a potential deadlock when some back-to-back phonebook loading
requests are received.

Consider that PB has been loaded (Sim card supports Global, do not
support Local), and adn cache is not null.

ThreadA is loading PB and acquiring synchronized(mLock). And right now
ThreadB intends to load PB, In this situation, ThreadB will be blocked
and waits for mLock release from ThreadA.

ThreadA:
Because of the adn cache is not null, the Global will return immediately,
The Local still need to send SIM_IO to read PBR, it will return quickly
since sim card does not support local PB. Main Thread will send a Message
to handler of IccPhoneBookInterfaceManager to notify. And ThreadA return
to wait() which will release synchronized mLock. so ThreadB can acquire
mLock and try to load PB.

ThreadB:
Same as ThreadB. Pls be Note, ThreadB is acquiring mLock in
IccPhoneBookInterfaceManager, when ThreadB is sending SIM_IO to load
Local PB, and will be waiting on mLock of UsimPhoneBookManager, the lock
will be notify after main thread send message back.

When the message to handler of IccPhoneBookInterfaceManager is earlier
than the message to Local PBR SIM_IO, it will result in deadlock. because
main thread will try to acquire mLock in IccPhoneBookInterfaceManager,
which is holding by ThreadB, however ThreadB is waiting the SIM_IO back
message by main thread.

CRs-Fixed: 727883

Change-Id: I4c2c0f7a817c5a6fb4be0d59197986e5f506b450