Better automated fallbacks in MccTable
Rather than keep a very incomplete list of the main language spoken in every
country, use CLDR data instead.

Also switch to the persist.sys.override_mcc system property for testing this
code.

Also bring the list of country codes up to date with the 2012 edition of the
ITU list

Bug: 17664563
Change-Id: I609619e3ba6f1683afe0e1b755a72e7d93f39b3d