PduParser MSIM support (2/4).
PduParser needs to know if Content-Disposition part header is supported
for a specific carrier. Originally this is loaded directly from system
resource in PduParser. This won't work for multi-SIM, which requires
getting that config based on carrier. This change adds parameter to
indicate if this header is supported and requires caller to provider the
config.

b/18371228

Change-Id: I2394dbb9551280037ad52e6983e35fa1f26d86ca