Fix race condition during LTE/eHRPD -> 1X IRAT scenario.
Instead of retrying data calls immediately due to NW_TYPE_CHANGED,
post an event to the handler so that DISCONNECT_DONE events queued
up can be handled before SETUP_DATA_CALL is attempted.
This is to avoid out of sync DISCONNECT events corrupting the
ApnContext states.
Ignore DISCONNECT event received when ApnContext is in CONNECTING
state as it could be a previous disconnect arriving at DCT late.
For ApnContext in CONNECTING state, only DATA_SETUP_COMPLETE is
valid.

Change-Id: I551cf14176953c89d2da93b16473d36c30e0ad44
CRs-Fixed: 766449, 779622