Telephony: Handle PS DETACH failure during DDS SWITCH
In DSDS, PS DETACH can fail if there is an ongoing voice call on the same sub.
This scenario is triggered when sub1 is DDS and is in a voice call. At this time
there is an MMS on sub2. DdsScheduler goes into DdsSwitchState and initiates an
on demand DDS switch from sub1 to sub2. This involves doing a PS DETACH on sub 1
followed by PS ATTACH on sub2.
If PS DETACH itself does not go through, data connection on sub1 will not be torn
down. As a result, DdsScheduler will be stuck in DdsSwitch state unable to process
further DDS switch requests.
Handle this by introducing PS DETACH retries for 6 times and notifying DdsScheduler
of the failure so that it goes back to IdleState to process further requests

CRs-Fixed: 759265
Change-Id: Ia426a51167be007847b1edde444eb154bb536e5a