In SipPhoneBase, do not notify call state changes
For Telephony phones, we'd like to notify call state changes in this
scenario, but for SIP calls we aren't concerned with this.

Tested: Added a custom listener to confirm that onCallStateChanged
previously was triggered on PhoneStateListener for specific subIds
and SIP calls. After this change, the callback is no longer invoked
for SIP calls, but still is for specific subIds.

Bug: 17612140

Change-Id: I9e4c6282fd53135580203c0b222782c23add5c42