IMS: Add support for SRVCC with more than 3 participants.
There is a max limit of 7 connections in GsmCallTracker. If RIL
reports call id greater than 7, it will be ignored while polling in
call tracker. Because of this, in case of SRVCC with more than 3
participants, call ids greater than 7 are getting dropped from UI.
Fix this by increasing the max connections limit to accomodate call
ids after SRVCC.

Change-Id: Id59712570e8a6e1675d1dd6a2cf069c90352abcd
CRs-Fixed: 736050