IMS: Update Connect time for dialing/incoming SRVCC calls
Connect time (mConnectTimeReal) will be updated only for ACTIVE calls
and it's not updated for calls which are transferred from DIALING/ALERTING/
INCOMING/WAITING to ACTIVE as part of SRVCC. Due to this once call is
ended call duration will not be updated properly in connection.

Update Connect time properly for DIALING/ALERTING/INCOMING/WAITING to
ACTIVE SRVCC calls.

Change-Id: I836dfcfd546ec55da44e69ff6cb90736521c001a
CRs-Fixed: 804727