Fix IccId query after Modem reset.
Currently if modem reset/SSR happens while iccid query
is in going on, then icc id query gets aborted and the
icc id value is not updated. This results in no subscription
activation.

Fix: If the old card state and new card state is present
and the iccid is ICCID_STRING_FOR_NO_SIM then requery for iccid.

Change-Id: I89d98bf33e67d11168c97c849d5db68f9b8b3ec3
CRs-Fixed: 811251.