Fix DDS switch issue when data call is attempted in OOS.
When setup data call fails in data out of service scenario,
DataConnection transitions immediately to Inactive state.
This disconnect event when ApnContext is in CONNECTINNG state
is not handled correctly, leaving ApnContext in CONNECTING
state forever. Fix this by handling DISCONNECT event even
when ApnContext is in CONNECTING state.

Change-Id: I754c4ed51729afad0389a451cf319bfa4217268c
CRs-Fixed: 814172