Fix Dds switch issue in case of active voice calls
When there is a MT call triggered, and before the MT call indication
is reached to Telephony, if there is a Dds switch started, it will
be rejected as MT call is active at modem.
This failed Dds switch is retried multiple times at Telephony and
once the maximum retries are exhausted, there will be no PS service
on any of the subs.

Add a check for voice call and stop retry in case of Dds switch failure.
Retry Dds switch on voice call end.

Change-Id: Iff86a8041530b11336a3ae4abce5de19703c878f
CRs-Fixed: 838955