MSIM: reset the subStatus during SSR
During SSR if SIM initialization is completed after
STACK_READY_EVENT is received would result in SIM not getting
provisioned.
In order to avoid such race condition reset the
subStatus to INIT and set the iccid to null during SSR
so that subscriptions are provisioned irrespective of
the delay in sim initialization.

Change-Id: I72a5b0d79519df2ef8f44bdadae8dc12789b925c
CRs-Fixed: 857550