Correctly remove pending network requests from pending list
If TelephonyNetworkFactory receives a network request when sub info
is not ready, it is put into pending list. Some times when release
network is called, sub info may be ready but the request may not have
been executed. In such cases, the request is not removed from the
pending list and causes extra network request handling and invalid
reference count for apn context.
To fix this, always check if a network request is in the pending list
and remove it if found. If the network request is present in the
pending list it means it was never executed.
Also remove the redundant sub info record broadcast receiver as
OnSubscriptionsChangedListener provides the same functionality.

Change-Id: I5e67cbfd77e84af55094754352eee8de45291d33
CRs-Fixed: 842725