Fix to limit correct boundary of unicode long messages.
Issue: The last 4 byte length character of the message part
of Unicode long message is getting truncated and shown as
unknown character on receiving.

Fix: If last character of a message part is 4 byte length
and it's message part length exceeds maximum sms length,
adjust the length of the message part length to
previous boundary position of the message part.

Change-Id: Ic274c8aa4b8d5e56d519f86ca0b884ef2bb592fd