Fix unwanted Dds switch issue for Mms
While voice call is active, if there is a Mms on non Dds
sub, Dds switch is rejected. On voice call end, the failed
Dds switch is re-tried on non Dds sub. This is leading to
permanent Dds switch to non Dds sub.

Add a check to ignore registration for voice call end if
Mms is in progress.

Change-Id: I0fa4f3a0320ae19c61923dbdedebfb7f5b67526d
CRs-Fixed: 900752