If only one sim in DSDS map it to primary stack.
In case of single sim is present in DSDS mode, then map
it to primary stack by doing flex mapping.
If sub id based network mode is not available, it means
that there is no sim card inserted in that slot.
Set the network mode to GSM_ONLY and evaluate flex
mapping as required.

Change-Id: Ic03e9bd9702d707a134f6c58704391979df1926a
CRs-Fixed: 912738