Fix Dds switch retry issue during voice call
On a multisim device, with Dds on Sub1, if an emergency call is done
on Sub1 and sim is removed from Slot1, Dds switch is tried to Sub2.
But as voice call is active, Dds switch to Sub2 is rejected.
Retry is being done on last known Dds sub for all Dds switch failures.
This is leading to Dds switch retry on Sub1 multiple times.

Fix is added to ignore Dds switch retry in case of active voice call
and Dds switch retry will be done on the active sub available.

Change-Id: I2f06ebd458e70b08741894df76a84bc47077e8c8
CRs-Fixed: 906328