Fix to bring up ims pdn when dds switched back to back.
ATEL will not send PS attach on a subscription which is already in
attached state.But RIL expects ATEL to send ps attach request on the
subscription Which is selected as new dds even if it is already has
PS attached.
If the new dds subscription is already in attached state then
transition to IDLE state in order to send PS attach request to RIL.
When data registers on IWLAN move the state machine to attached state.

Change-Id: I8f5d9022c2d12d5c0b3384c3a96d5170169f58e2
CRs-Fixed: 1007581